flask_api/requirements.txt
Flask>=3.0.0
Flask-CORS>=4.0.0
python-dotenv>=1.0.0
requests>=2.31.0
qdrant-client>=1.7.0
openai>=1.3.0
numpy>=1.24.0
pydantic>=2.5.0
gunicorn>=21.2.0 
pandas
openpyxl 